Moved some tests into subfolder
