+++ test/Criterion/Test/Helper/HelperTest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Criterion\Test\Helper;
+
+class HelperTest extends \Criterion\Test\TestCase
+{
+
+    public function testInit()
+    {
+        $helper = new \Criterion\Helper();
+        $this->assertTrue($helper instanceof \Criterion\Helper);
+    }
+}
