--- test/Criterion/Test/HelperTest.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Criterion\Test;
-
-class HelperTest extends TestCase
-{
-
-    public function testInit()
-    {
-        $helper = new \Criterion\Helper();
-        $this->assertTrue(is_object($helper));
-    }
-}
